Apply reviews (Merge applicant_status to applicant, clear questions)
