--- src/app/applicant/api/router/applicant_status_router.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-const applicantStatusController = require('../controller/applicant_status_controller.js');
-
-router.patch('/:id', applicantStatusController.changeApplicantStatus);
-router.patch('/', applicantStatusController.changeApplicantListStatus);
-
-module.exports = router;
-
